backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
umap-learn==0.5.6
numpy==2.1.1
pydantic==2.9.2

